# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Reviewed AlKabir Islamic Challenge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/feedback:
    get:
      operationId: listMyFeedback
      tags: [profile]
      responses:
        "200":
          description: Feedback submitted by the member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createMyFeedback
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackInput" }
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feedback" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/beta/feedback:
    get:
      operationId: listBetaFeedback
      tags: [admin]
      parameters:
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/FeedbackStatus" }
        - in: query
          name: kind
          schema: { $ref: "#/components/schemas/FeedbackKind" }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Feedback queue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackList" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/beta/feedback/{feedbackId}:
    patch:
      operationId: moderateBetaFeedback
      tags: [admin]
      parameters:
        - in: path
          name: feedbackId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackModerationInput" }
      responses:
        "200":
          description: Moderated feedback
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feedback" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /admin/beta/audit:
    get:
      operationId: listBetaAudit
      tags: [admin]
      parameters:
        - in: query
          name: action
          schema: { type: string, maxLength: 100 }
        - in: query
          name: entityType
          schema: { type: string, maxLength: 100 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Safe operator audit events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditList" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /auth/link-guest-progress:
    post:
      operationId: linkGuestProgress
      tags: [auth]
      summary: Explicitly link the current anonymous progress to the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkGuestProgressRequest"
      responses:
        "200":
          description: Guest progress was linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkGuestProgressResult"
        "400":
          description: Explicit confirmation is missing
        "401":
          description: Authentication is required
        "403":
          description: Origin or CSRF validation failed
        "410":
          description: Anonymous session expired
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    FeedbackKind:
      type: string
      enum: [question_accuracy, technical, accessibility, general]
    FeedbackStatus:
      type: string
      enum: [open, in_review, resolved, dismissed]
    FeedbackInput:
      type: object
      required: [kind, message]
      properties:
        kind: { $ref: "#/components/schemas/FeedbackKind" }
        message: { type: string, minLength: 1, maxLength: 5000 }
        questionId: { type: string, format: uuid }
        questionVersionId: { type: string, format: uuid }
    FeedbackModerationInput:
      type: object
      required: [status, expectedStatus]
      properties:
        status: { $ref: "#/components/schemas/FeedbackStatus" }
        expectedStatus: { $ref: "#/components/schemas/FeedbackStatus" }
        resolutionNote: { type: string, maxLength: 2000, nullable: true }
    Feedback:
      type: object
      required: [id, kind, message, status, questionId, questionVersionId, resolutionNote, resolvedAt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/FeedbackKind" }
        message: { type: string }
        status: { $ref: "#/components/schemas/FeedbackStatus" }
        questionId: { type: string, format: uuid, nullable: true }
        questionVersionId: { type: string, format: uuid, nullable: true }
        resolutionNote: { type: string, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FeedbackList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Feedback" }
        total: { type: integer, minimum: 0 }
...
    AuditEvent:
      type: object
      required: [id, actorId, action, entityType, entityId, metadata, createdAt]
      properties:
        id: { type: string, format: uuid }
        actorId: { type: string, maxLength: 256 }
        action: { type: string, maxLength: 100 }
        entityType: { type: string, maxLength: 100 }
        entityId: { type: string, maxLength: 256 }
        metadata:
          type: object
          additionalProperties:
            oneOf: [{ type: string }, { type: number }, { type: boolean }, { type: "null" }]
        createdAt: { type: string, format: date-time }
    AuditList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditEvent" }
        total: { type: integer, minimum: 0 }
...
    LinkGuestProgressRequest:
      type: object
      properties:
        confirm:
          type: boolean
      required: [confirm]
    LinkGuestProgressResult:
      type: object
      properties:
        linked:
          type: boolean
        linkedAttemptCount:
          type: integer
          minimum: 0
      required: [linked, linkedAttemptCount]
...